[REFACTOR] change MDe implementation to l10n_br_nfe
