+++ l10n_br_fiscal_dfe/models/__init__.py
@@ -1,5 +1,4 @@
 from . import dfe
-from . import mde
 from . import document
 from . import attachment
 from . import res_company
