+++ l10n_br_fiscal_dfe/security/ir.model.access.csv
@@ -1,5 +1,3 @@
 "id","name","model_id:id","group_id:id","perm_read","perm_write","perm_create","perm_unlink"
 "l10n_br_fiscal_dfe_user","Consut DFe for User","model_l10n_br_fiscal_dfe","l10n_br_fiscal.group_user",1,1,1,0
 "l10n_br_fiscal_dfe_manager","Consut DFe for Manager","model_l10n_br_fiscal_dfe","l10n_br_fiscal.group_manager",1,1,1,1
-"l10n_br_fiscal_mde_user","MDe for User","model_l10n_br_fiscal_mde","l10n_br_fiscal.group_user",1,1,1,0
-"l10n_br_fiscal_mde_manager","MDe for Manager","model_l10n_br_fiscal_mde","l10n_br_fiscal.group_manager",1,1,1,1
