+++ l10n_br_fiscal_dfe/tests/__init__.py
@@ -1,2 +1,1 @@
 from . import test_dfe
-from . import test_mde
