+++ l10n_br_fiscal_dfe/tests/test_dfe.py
@@ -10,7 +10,6 @@ from nfelib.v4_00 import retDistDFeInt
 
 from odoo.tests.common import SavepointCase
 
-from ..models.mde import MDe
 from ..tools import utils
 
 # flake8: noqa: B950
@@ -89,29 +88,6 @@ class TestDFe(SavepointCase):
         self.dfe_id.search_documents()
         self.assertEqual(self.dfe_id.last_nsu, utils.format_nsu("201"))
 
-        self.assertEqual(self.dfe_id.mde_ids[-1].nsu, self.dfe_id.last_nsu)
-
-        mde1, mde2 = self.dfe_id.mde_ids
-        self.assertEqual(mde1.company_id, self.dfe_id.company_id)
-        self.assertEqual(mde1.key, "31201010588201000105550010038421171838422178")
-        self.assertEqual(mde1.emitter, "ZAP GRAFICA E EDITORA EIRELI")
-        self.assertEqual(mde1.cnpj_cpf, "10.588.201/0001-05")
-        self.assertEqual(mde1.state, "pendente")
-
-        attachment_1 = self.env["ir.attachment"].search([("res_id", "=", mde1.id)])
-        self.assertTrue(attachment_1)
-
-        self.assertEqual(mde2.company_id, self.dfe_id.company_id)
-        self.assertEqual(mde2.key, "35200159594315000157550010000000012062777161")
-        self.assertEqual(
-            mde2.partner_id, self.env.ref("l10n_br_base.simples_nacional_partner")
-        )
-        self.assertEqual(mde2.cnpj_cpf, "59.594.315/0001-57")
-        self.assertEqual(mde2.state, "pendente")
-
-        attachment_2 = self.env["ir.attachment"].search([("res_id", "=", mde2.id)])
-        self.assertTrue(attachment_2)
-
     def test_search_dfe_error(self):
         with mock.patch.object(
             DocumentoElectronicoAdapter,
@@ -120,7 +96,6 @@ class TestDFe(SavepointCase):
         ):
             self.dfe_id.search_documents()
             self.assertEqual(self.dfe_id.last_nsu, "000000000000000")
-            self.assertFalse(self.dfe_id.mde_ids)
 
         with mock.patch.object(
             DocumentoElectronicoAdapter,
@@ -129,7 +104,6 @@ class TestDFe(SavepointCase):
         ):
             self.dfe_id.search_documents()
             self.assertEqual(self.dfe_id.last_nsu, "000000000000000")
-            self.assertFalse(self.dfe_id.mde_ids)
 
         with mock.patch.object(
             DocumentoElectronicoAdapter,
@@ -137,46 +111,8 @@ class TestDFe(SavepointCase):
             side_effect=KeyError("foo"),
         ):
             self.dfe_id.search_documents()
-            self.assertFalse(self.dfe_id.mde_ids)
 
-    @mock.patch.object(
-        DocumentoElectronicoAdapter,
-        "_post",
-        side_effect=mocked_post_success_single,
-    )
-    @mock.patch.object(MDe, "action_ciencia_emissao", return_value=None)
-    def test_import_documents(self, _mock_post, _mock_ciencia):
-        self.dfe_id.search_documents()
-        self.dfe_id.import_documents()
-
-        with mock.patch.object(
-            DocumentoElectronicoAdapter,
-            "_post",
-            side_effect=mocked_post_error_status_code,
-        ):
-            xml = self.dfe_id._download_document("dummy")
-            self.assertIsNone(xml)
-
-    def test_create_mde(self):
-        mde = self.dfe_id._create_mde_from_schema("dummy_v1.0", False)
-        self.assertIsNone(mde)
-
-        mde_id = self.env["l10n_br_fiscal.mde"].create({"key": "123456789"})
-
-        mock_resNFe = mock.MagicMock(spec=["chNFe"])
-        mock_resNFe.chNFe = "123456789"
-        resnfe_mde_id = self.dfe_id._create_mde_from_schema("resNFe_v1.0", mock_resNFe)
-        self.assertEqual(resnfe_mde_id, mde_id)
-
-        mock_procNFe = mock.MagicMock(spec=["protNFe"])
-        mock_procNFe.protNFe.infProt.chNFe = "123456789"
-        procnfe_mde_id = self.dfe_id._create_mde_from_schema(
-            "procNFe_v1.0", mock_procNFe
-        )
-        self.assertEqual(procnfe_mde_id, mde_id)
-
-    @mock.patch.object(MDe, "action_ciencia_emissao", return_value=None)
-    def test_cron_search_documents(self, _mock):
+    def test_cron_search_documents(self):
         self.dfe_id.use_cron = True
 
         with mock.patch.object(
@@ -185,7 +121,7 @@ class TestDFe(SavepointCase):
             side_effect=mocked_post_error_status_code,
         ):
             self.dfe_id._cron_search_documents()
-            self.assertFalse(self.dfe_id.mde_ids)
+            self.assertEqual(self.dfe_id.last_nsu, "000000000000000")
 
         with mock.patch.object(
             DocumentoElectronicoAdapter,
@@ -193,7 +129,7 @@ class TestDFe(SavepointCase):
             side_effect=mocked_post_success_multiple,
         ):
             self.dfe_id._cron_search_documents()
-            self.assertEqual(len(self.dfe_id.mde_ids), 2)
+            self.assertEqual(self.dfe_id.last_nsu, "000000000000201")
 
     def test_utils(self):
         nsu_formatted = utils.format_nsu("100")
