+++ l10n_br_nfe/__manifest__.py
@@ -28,6 +28,8 @@
         "views/res_company_view.xml",
         "views/nfe_document_view.xml",
         "views/res_config_settings_view.xml",
+        "views/mde/mde_views.xml",
+        "views/dfe/dfe_views.xml",
         # Report
         "report/reports.xml",
         "report/danfe_nfce.xml",
