+++ l10n_br_nfe/constants/mde.py
