+++ l10n_br_nfe/models/__init__.py
@@ -19,3 +19,4 @@ from . import res_config_settings
 from . import cfop
 from . import invalidate_number
 from . import dfe
+from . import mde
