+++ l10n_br_nfe/security/ir.model.access.csv
@@ -1,0 +1,5 @@
+"id","name","model_id:id","group_id:id","perm_read","perm_write","perm_create","perm_unlink"
+"l10n_br_fiscal_import_wizard_user",l10n_br_nfe_import_xml,model_l10n_br_nfe_import_xml,base.group_user,1,1,1,1
+"l10n_br_fiscal_import_products_wizard_user",l10n_br_nfe_import_xml_products,model_l10n_br_nfe_import_xml_products,base.group_user,1,1,1,1
+"l10n_br_fiscal_mde_user","MDe for User","model_l10n_br_fiscal_mde","l10n_br_fiscal.group_user",1,1,1,0
+"l10n_br_fiscal_mde_manager","MDe for Manager","model_l10n_br_fiscal_mde","l10n_br_fiscal.group_manager",1,1,1,1
