+++ l10n_br_nfe/tests/__init__.py
@@ -8,3 +8,4 @@ from . import test_nfe_webservices
 from . import test_nfe_xml_validation
 from . import test_res_partner
 from . import test_nfe_dfe
+from . import test_nfe_mde
